requirements.txt
python-telegram-bot==13.7
urllib3==1.26.15
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-generativeai==0.3.2
